Deselect snapshot on filter change
